moved static parser code from DBEntity to Importer
